Fix explorer_passport: cast the week offset to int
Defect found during Stage 5 validation, against acceptance §0 (migrations applied
and schema verified). Fix only; no other change.

explorer_passport() threw 42883 "operator does not exist: date - bigint" on every
call, so the Passport rendered empty on every profile. row_number() returns bigint
and `date - bigint` has no operator — only `date - integer` — so the streak
query's week offset needs an explicit cast.

The surrounding streak logic is correct and unchanged: because week-truncated
dates differ only in multiples of 7, a row can never re-align with the required
offset after a gap, so counting matching rows gives exactly the leading
consecutive run.

Caught by a post-apply verification pass against the live schema rather than by
the app, because the panel fails soft by design (retry:false, renders nothing) —
it would have looked like "no data yet" instead of an error.

Re-run the file in the SQL editor; it is idempotent.

+++ supabase/migrations/20260801130000_explorer_passport.sql
@@ -103,9 +103,12 @@ begin
   ), ranked as (
     select w, row_number() over (order by w desc) as rn from weeks
   )
+  -- row_number() is bigint and `date - bigint` has no operator (only date - integer),
+  -- so the week offset must be cast explicitly. Without it the whole function throws
+  -- 42883 and every Passport renders empty.
   select count(*) into v_streak from ranked
    where (select max(w) from weeks) >= date_trunc('week', now())::date - 7
-     and w = (select max(w) from weeks) - ((rn - 1) * 7);
+     and w = (select max(w) from weeks) - (((rn - 1) * 7)::int);
 
   -- Level rewards the things Purday wants more of: verified presence, blueprints
   -- other people actually copied, real outings. Square-rooted so it climbs steadily
